requirements.txt
fastapi[standard]
pydantic
uvicorn[standard]
python-dotenv
pydantic-settings
sqlalchemy
psycopg2